requirements.txt
functions-framework==3.*
torch
diffusers
transformers==4.22.2
accelerate==0.13.2
ftfy==6.1.1
spacy==3.4.3
google-cloud-storage